# specs/001-tv-channel-programming/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SmartTunarr API
  version: 1.0.0
  description: Intelligent TV Channel Programming System API
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /history:
    get:
      tags: [History]
      summary: List operation history
      operationId: listHistory
      parameters:
        - name: channel_id
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            type: string
            enum: [programming, scoring]
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: History entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryList'
  /history/{entryId}:
    get:
      tags: [History]
      summary: Get history entry details
      operationId: getHistoryEntry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: History entry details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryEntry'
        '404':
          $ref: '#/components/responses/NotFound'

  # AI endpoints
  /ai/generate-profile:
    post:
      tags: [AI]
      summary: Generate profile from natural language description
      operationId: generateProfileFromPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIPromptRequest'
      responses:
        '200':
          description: Generated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIGeneratedProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ProfileImport:
      type: object
      required: [profile_name, version, libraries, time_blocks, mandatory_forbidden_criteria, scoring_weights]
      properties:
        profile_name:
          type: string
        version:
          type: string
        libraries:
          type: array
          items:
            $ref: '#/components/schemas/LibraryRef'
        time_blocks:
          type: array
          items:
            $ref: '#/components/schemas/TimeBlock'
        mandatory_forbidden_criteria:
          $ref: '#/components/schemas/Criteria'
        strategies:
          type: array
          items:
            $ref: '#/components/schemas/Strategy'
        scoring_weights:
          $ref: '#/components/schemas/ScoringWeights'
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        warnings:
          type: array
          items:
            type: string
    ValidationError:
      type: object
      properties:
        path:
          type: string
        message:
          type: string
        expected:
          type: string
    LibraryRef:
      type: object
      required: [name, type, plex_id]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [movie, show, trailer]
        plex_id:
          type: string
    TimeBlock:
      type: object
      required: [name, start_time, end_time]
      properties:
        name:
          type: string
        start_time:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
        end_time:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
        criteria:
          $ref: '#/components/schemas/Criteria'
    Criteria:
      type: object
      properties:
        genre_criteria:
          $ref: '#/components/schemas/GenreCriteria'
        age_criteria:
          $ref: '#/components/schemas/AgeCriteria'
        rating_criteria:
          $ref: '#/components/schemas/RatingCriteria'
        content_safety:
          $ref: '#/components/schemas/ContentSafety'
    GenreCriteria:
      type: object
      properties:
        mandatory_genres:
          type: object
          properties:
            required:
              type: boolean
            genres:
              type: array
              items:
                type: string
        forbidden_genres:
          type: object
          properties:
            enabled:
              type: boolean
            genres:
              type: array
              items:
                type: string
        preferred_genres:
          type: array
          items:
            type: string
    AgeCriteria:
      type: object
      properties:
        mandatory_age:
          type: object
          properties:
            required:
              type: boolean
            allowed_ratings:
              type: array
              items:
                type: string
        forbidden_age:
          type: object
          properties:
            enabled:
              type: boolean
            forbidden_ratings:
              type: array
              items:
                type: string
    RatingCriteria:
      type: object
      properties:
        mandatory_rating:
          type: object
          properties:
            required:
              type: boolean
            min_tmdb_rating:
              type: number
            min_vote_count:
              type: integer
    ContentSafety:
      type: object
      properties:
        forbidden_keywords:
          type: object
          properties:
            enabled:
              type: boolean
            dangerous_keywords:
              type: array
              items:
                type: string
        mandatory_keywords:
          type: object
          properties:
            required:
              type: boolean
            safe_keywords:
              type: array
              items:
                type: string
    Strategy:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [movie_sequence, trailer_insert, genre_block, rating_progression]
        count:
          type: integer
        after:
          type: integer
        genres:
          type: array
          items:
            type: string
    ScoringWeights:
      type: object
      properties:
        type:
          type: integer
          minimum: 0
          maximum: 100
        duration:
          type: integer
          minimum: 0
          maximum: 100
        genre:
          type: integer
          minimum: 0
          maximum: 100
        timing:
          type: integer
          minimum: 0
          maximum: 100
        strategy:
          type: integer
          minimum: 0
          maximum: 100
        age:
          type: integer
          minimum: 0
          maximum: 100
        rating:
          type: integer
          minimum: 0
          maximum: 100
        filters:
          type: integer
          minimum: 0
          maximum: 100
        bonus:
          type: integer
          minimum: 0
          maximum: 100

    # Channel schemas
...
    HistoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HistorySummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    HistorySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [programming, scoring]
        channel_name:
          type: string
        profile_name:
          type: string
        status:
          type: string
          enum: [running, success, failed]
        best_score:
          type: number
        started_at:
          type: string
          format: date-time
    HistoryEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [programming, scoring]
        channel_id:
          type: string
          format: uuid
        channel_name:
          type: string
        profile_id:
          type: string
          format: uuid
        profile_name:
          type: string
        status:
          type: string
          enum: [running, success, failed]
        iterations:
          type: integer
        best_score:
          type: number
        result_summary:
          type: object
        error_message:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    # AI schemas
    AIPromptRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 10
          maxLength: 2000
        model:
          type: string
    AIGeneratedProfile:
      type: object
      properties:
        profile:
          $ref: '#/components/schemas/ProfileImport'
        validation:
          $ref: '#/components/schemas/ValidationResult'
        prompt_used:
          type: string
        model_used:
          type: string
        generated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object